+++ packages/origin-backend/src/api/actions/IActions.ts
@@ -1,7 +1,8 @@
-import { Request, Response } from "express";
+import { Request, Response } from 'express';
 
 export interface IActions {
     get: (req: Request, res: Response) => Promise<void>;
     post: (req: Request, res: Response) => Promise<void>;
     delete: (req: Request, res: Response) => Promise<void>;
+    put?: (req: Request, res: Response) => Promise<void>;
 }