+++ packages/origin-backend-client/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.2](https://github.com/energywebfoundation/origin/compare/@energyweb/origin-backend-client@2.2.1...@energyweb/origin-backend-client@2.2.2) (2020-01-07)
+
+**Note:** Version bump only for package @energyweb/origin-backend-client
+
+
+
+
+
 ## [2.2.1](https://github.com/energywebfoundation/origin/compare/@energyweb/origin-backend-client@2.2.0...@energyweb/origin-backend-client@2.2.1) (2019-12-20)